backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
djoser
python-dotenv